Add minimal shitty yaml parser based on encoding/json
I can't believe I'm saying this - but I want to write yaml instead of json when
configuring apps. This implements a yaml parser good enough for my needs in a
few hundred lines - because why not. I probably shouldn't call it yaml - that
spec is crazy - and this doesn't do anything more complex than json, which has
a much more approachable spec.