Increase dirty version timestamp granularity
the idea behind adding a timestamp is to prevent $(k version) == $(k version)
and thus force reinstall of the binary on the server for dirty builds - because
we have no idea what changed, we just assume something changed. second
granularity isn't enough - nano should be. we could also use random but that's
more lines