k-notify: Do not mark unit as failed if notify fails
We want to (attempt to) send a notification everytime a unit fails -
independent of whether sending a notification failed before. And we don't want
failures to send notifications to affect the systemctl status of k. By
prefixing the ExecStart command with `-` we prevent the instantiated notify
unit from being marked as failed when `k notify` fails. This removes useless
noise from `systemctl --failed` without having to run the equivalent of
`systemctl reset-failed`.