config: Pass EnvFile to services to support secrets
We can't store secrets in unit file `Environment=` lines because systemd
considers unit files public and ignores file permissions when running
`systemctl cat $service`. Using `EnvironmentFile=` allows us to use env
variables for secrets and make use of file permissions to keep them secret.

The systemd folks don't think env variables are the right way to pass
secrets and instead implemented credentials [1]. They're reasoning is [1]:

    Traditionally, data of this nature has often been provided to services via
    environment variables (which is problematic because by default they are
    inherited down the process tree, have size limitations, and issues with
    binary data)

I agree with their reasoning, but I don't think the value/cost ratio of
implementing credentials (and complicating the interface of my personal
services) is worth it - the only drawback that I see applying to my services is
the leaking of secrets to child processes - and that's a risk I'm willing to
accept when the cost of preventing it is so high / annoying. Let's see if I'll
regret this.

---

There's quite a few other changes to the config code in support of making it
easier to test. Also mergeUnits was adopted to allow merging values by
specifying the default as an array (to support having a default EnvFile while
still allowing the service to specfiy their own)

[1] https://systemd.io/CREDENTIALS/